split the js file into multiple modules
